+++ Tests/Unit/Controller/UserWithoutAutologinControllerTest.php
@@ -19,10 +19,13 @@ use OliverKlee\Onetimeaccount\Validation\UserValidator;
 use PHPUnit\Framework\MockObject\MockObject;
 use TYPO3\CMS\Core\Context\Context;
 use TYPO3\CMS\Core\Http\HtmlResponse;
+use TYPO3\CMS\Core\Http\RedirectResponse;
+use TYPO3\CMS\Core\Information\Typo3Version;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
 use TYPO3\CMS\Extbase\Mvc\Controller\ActionController;
 use TYPO3\CMS\Extbase\Mvc\Controller\Argument as ExtbaseArgument;
 use TYPO3\CMS\Extbase\Mvc\Controller\Arguments;
+use TYPO3\CMS\Extbase\Mvc\Exception\StopActionException;
 use TYPO3\CMS\Extbase\Mvc\Request;
 use TYPO3\CMS\Extbase\Persistence\ObjectStorage;
 use TYPO3\CMS\Extbase\Persistence\QueryResultInterface;
@@ -177,6 +180,23 @@ final class UserWithoutAutologinControllerTest extends UnitTestCase
     private function resetRequestData(): void
     {
         GeneralUtility::flushInternalRuntimeCaches();
+    }
+
+    /**
+     * @param non-empty-string $uri
+     */
+    private function mockRedirectToUri(string $uri): void
+    {
+        if ((new Typo3Version())->getMajorVersion() < 12) {
+            $this->subject->expects(self::once())->method('redirectToUri')
+                ->with($uri)
+                ->willThrowException(new StopActionException('redirectToUri', 1476045828));
+            $this->expectException(StopActionException::class);
+        } else {
+            $redirectResponse = $this->createStub(RedirectResponse::class);
+            $this->subject->expects(self::once())->method('redirectToUri')->with($uri)
+                ->willReturn($redirectResponse);
+        }
     }
 
     /**
@@ -1034,9 +1054,14 @@ final class UserWithoutAutologinControllerTest extends UnitTestCase
         $this->credentialsGeneratorMock->method('generateAndSetPasswordForUser')
             ->with(self::anything())
             ->willReturn('');
-        $this->subject->expects(self::once())->method('redirectToUri')->with($redirectUrl);
+        $this->mockRedirectToUri($redirectUrl);
 
-        $this->subject->createAction(new FrontendUser());
+        if ((new Typo3Version())->getMajorVersion() < 12) {
+            $this->subject->createAction(new FrontendUser());
+        } else {
+            $result = $this->subject->createAction(new FrontendUser());
+            self::assertInstanceOf(RedirectResponse::class, $result);
+        }
     }
 
     /**